Updated Gibbs sampler for Bayesian PMF
If we factor R = UV^T, with appropriate priors on U and V, then I have a Gibbs sampler for U and its priors given V and a Gibbs sampler for V and its priors given U.  Currently, the synthetic dataset is full - there is zero sparsity.  Some experimentation indicates that the levels of sparsity present in the data could be a problem.