Wrote script to generate test/train sets
Takes input from ratings matrix, generates a training set, a test set, and cluster matrices for each.  In this implementation, the training and test sets are randomized on each user-item pair, so that items can appear in both the training and test sets.  This is necessary for BPMF to make sense.