Working on occupancy model
Finished some of the occupancy model --- the entire model works if we assume that the W_nm are data.  We might be able to apply this model to the data taking the ratings as the W_nm, if we just care about which movies are rated and not about which movies are watched and rated.