Added W_nm updating
The model is a lot worse if we have to learn the true W_nm --- don't learn any latent structure correctly.  May need to go back to the drawing board -- either to make the cluster stucture more structured (that is, take the clusters as given) or to make the probabilities more principled.